requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-import-export==3.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
xlrd==2.0.1
xlwt==1.3.0
